# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mistral.rs",
    "description": "Fast, flexible LLM inference.",
    "contact": {
      "name": "Eric Buehler"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.9.4"
  },
  "paths": {
...
    "/v1/systemone": {
      "post": {
        "tags": [
          "System One"
        ],
        "operationId": "system_one",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Typed decisions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid question, input, or model capability"
          },
          "404": {
            "description": "Model not found"
          },
          "500": {
            "description": "Inference failed"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DecisionAnswer": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "noul",
              "type"
            ],
            "properties": {
              "noul": {
                "type": "number",
                "format": "double"
              },
              "type": {
                "type": "string",
                "enum": [
                  "noul"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "choice",
              "confidence",
              "probabilities",
              "type"
            ],
            "properties": {
              "choice": {
                "type": "string"
              },
              "confidence": {
                "type": "number",
                "format": "double"
              },
              "probabilities": {
                "type": "object",
                "additionalProperties": {
                  "type": "number",
                  "format": "double"
                },
                "propertyNames": {
                  "type": "string"
                }
              },
              "type": {
                "type": "string",
                "enum": [
                  "choice"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "score",
              "confidence",
              "probabilities",
              "legend",
              "type"
            ],
            "properties": {
              "confidence": {
                "type": "number",
                "format": "double"
              },
              "legend": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                },
                "propertyNames": {
                  "type": "string"
                }
              },
              "probabilities": {
                "type": "object",
                "additionalProperties": {
                  "type": "number",
                  "format": "double"
                },
                "propertyNames": {
                  "type": "string"
                }
              },
              "score": {
                "type": "number",
                "format": "double"
              },
              "type": {
                "type": "string",
                "enum": [
                  "score"
                ]
              }
            }
          }
        ]
      },
...
      "DecisionQuestion": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "criteria": {
                "type": [
                  "object",
                  "null"
                ],
                "additionalProperties": {},
                "propertyNames": {
                  "type": "string"
                }
              },
              "instructions": {},
              "type": {
                "type": "string",
                "enum": [
                  "noul"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "criteria",
              "type"
            ],
            "properties": {
              "criteria": {
                "type": "object",
                "additionalProperties": {},
                "propertyNames": {
                  "type": "string"
                }
              },
              "instructions": {},
              "type": {
                "type": "string",
                "enum": [
                  "choice"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "criteria",
              "type"
            ],
            "properties": {
              "criteria": {
                "type": "array",
                "items": {}
              },
              "instructions": {},
              "type": {
                "type": "string",
                "enum": [
                  "score"
                ]
              }
            }
          }
        ]
      },
      "DecisionRequest": {
        "type": "object",
        "required": [
          "model",
          "state",
          "questions"
        ],
        "properties": {
          "model": {
            "type": "string"
          },
          "questions": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/DecisionQuestion"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "state": {},
          "temperature": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "DecisionResponse": {
        "type": "object",
        "required": [
          "model",
          "answers",
          "usage"
        ],
        "properties": {
          "answers": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/DecisionAnswer"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "model": {
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/DecisionUsage"
          }
        }
      },
      "DecisionUsage": {
        "type": "object",
        "required": [
          "input_tokens",
          "output_tokens",
          "billing_units"
        ],
        "properties": {
          "billing_units": {
            "type": "integer",
            "minimum": 0
          },
          "input_tokens": {
            "type": "integer",
            "minimum": 0
          },
          "output_tokens": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
    }
  },
...
}